+++ snowchains_core/examples/codeforces-retrieve-test-cases.rs
@@ -2,7 +2,6 @@ use cookie_store::CookieStore;
 use indicatif::ProgressDrawTarget;
 use snowchains_core::{
     color_spec,
-    testsuite::Match,
     web::{
         Codeforces, CodeforcesRetrieveSampleTestCasesCredentials, CookieStorage, ProblemsInContest,
         RetrieveTestCases, StatusCodeColor,
@@ -58,7 +57,6 @@ fn main() -> anyhow::Result<()> {
         credentials: CodeforcesRetrieveSampleTestCasesCredentials {
             username_and_password: &mut username_and_password(credentials),
         },
-        default_match: Match::Lines,
         full: None,
         cookie_storage: CookieStorage {
             cookie_store: CookieStore::default(),