+++ snowchains_core/examples/yukicoder-retrieve-test-cases.rs
@@ -1,7 +1,6 @@
 use indicatif::ProgressDrawTarget;
 use snowchains_core::{
     color_spec,
-    testsuite::Match,
     web::{
         RetrieveFullTestCases, RetrieveTestCases, StatusCodeColor, Yukicoder,
         YukicoderRetrieveFullTestCasesCredentials, YukicoderRetrieveTestCasesTargets,
@@ -72,7 +71,6 @@ fn main() -> anyhow::Result<()> {
             ),
         },
         credentials: (),
-        default_match: Match::Lines,
         full: if full {
             Some(RetrieveFullTestCases {
                 credentials: YukicoderRetrieveFullTestCasesCredentials {