+++ snowchains_core/tests/yukicoder-retrieve-test-cases.rs
@@ -1,9 +1,8 @@
 use insta::{assert_debug_snapshot, assert_json_snapshot};
 use maplit::btreeset;
 use reqwest::{Method, StatusCode};
-use snowchains_core::{
-    testsuite::Match,
-    web::{RetrieveTestCases, StatusCodeColor, Yukicoder, YukicoderRetrieveTestCasesTargets},
+use snowchains_core::web::{
+    RetrieveTestCases, StatusCodeColor, Yukicoder, YukicoderRetrieveTestCasesTargets,
 };
 use std::{
     fmt,
@@ -81,7 +80,6 @@ fn test(target: Target) -> anyhow::Result<()> {
             }
         },
         credentials: (),
-        default_match: Match::Lines,
         full: None,
         cookie_storage: (),
         timeout: TIMEOUT,